Conflicts:
	CMakeLists.txt
	src/application/client.cpp
	src/application/server.cpp
	src/entity/controllers/collisioncontroller.hpp
	src/entity/controllers/playercontroller.cpp
	src/entity/entitymessages.hpp
	src/rendering/chunkgeneratorqueue.hpp
	src/rendering/debugrenderer.hpp
	src/rendering/frustum.hpp
	src/rendering/marchingcubegenerator.hpp
	src/rendering/renderingmessages.hpp
	src/rendering/texture.hpp
	src/rendering/vbo.hpp
	src/script/interfaces/landscapeinterface.hpp
	src/utilities/fpscontroller.hpp
	src/utilities/geomath.hpp
	src/world/region.cpp
	src/world/region.hpp
	src/world/world.cpp
	src/world/world.hpp
	src/world/worldmessages.hpp
	src/world/worldprovider.cpp
	src/world/worldprovider.hpp
	src/world/worldsystem.cpp
	src/world/worldsystem.hpp
	test/world/coordinatetest.cpp
	test/world/misctest.cpp

+++ src/world/ranges.hpp
@@ -7,10 +7,10 @@ using IntRange = std::pair<int64_t, int64_t>;
 struct Ranges
 {
     Ranges();
-    static int32_t MAX;
-    static int32_t MIN;
+    static int64_t MAX;
+    static int64_t MIN;
     Ranges(const IntRange& x, const IntRange& y, const IntRange& z);
-    bool isWithin(const glm::ivec3& coordinate) const;
+    bool isWithin(const glm::i64vec3& coordinate) const;
     IntRange xRange;
     IntRange yRange;
     IntRange zRange;
